Refactored short alerts into one component
